initialize warehouse

+++ src/main/java/cz/cvut/junit/web/wrapper/init/BoxWrapper.java
@@ -10,7 +10,7 @@ import java.util.List;
  */
 public class BoxWrapper {
     @JsonProperty("number")
-    private int number;
+    private long number;
 
     @JsonProperty("type")
     private CoolingType type;
@@ -18,11 +18,11 @@ public class BoxWrapper {
     @JsonProperty("shelfs")
     private List<ShelfWrapper> shelfs;
 
-    public int getNumber() {
+    public long getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(long number) {
         this.number = number;
     }
